+++ src/inNodeJs.ts
@@ -1,0 +1,8 @@
+/**
+ * Determine if it is running on node.js
+ *
+ * @returns boolean
+ */
+const inNodeJs = typeof global !== 'undefined'
+
+export default inNodeJs