+++ src/index.default.ts
@@ -33,6 +33,7 @@ import isObject from './isObject'
 import isArray from './isArray'
 import isIterable from './isIterable'
 import inBrowser from './inBrowser'
+import inNodeJs from './inNodeJs'
 import windowSize from './windowSize'
 import getAppVersion from './getAppVersion'
 import appVersion from './appVersion'
@@ -204,6 +205,7 @@ export default {
 	isPlainObject,
 	isDarkMode,
 	inBrowser,
+	inNodeJs,
 	minus,
 	nextIndex,
 	nextVersion,